fixed a bug caused by instances being set directly to active in the simcloud

+++ algorithm/naive.go
@@ -49,8 +49,8 @@ func (n NaiveAlgorithm) Run(input autoscale.AlgorithmInput, startTime time.Time)
 				var duration int64
 				duration = 0
 				flav := "default"
-				if job.InstanceFlavour.Name != "" {
-					flav = job.InstanceFlavour.Name
+				if job.InstanceFlavour != "" {
+					flav = job.InstanceFlavour
 				}
 				cost := cloud.GetExpectedJobCost(job, flav, startTime)
 				if cost < lowestCost {
@@ -102,14 +102,25 @@ func (n NaiveAlgorithm) Run(input autoscale.AlgorithmInput, startTime time.Time)
 		})
 
 		runningJobs := 0
+		activeInstances := 0
+		for _, i := range instances {
+			if i.State == autoscale.ACTIVE {
+				activeInstances++
+			}
+		}
+
+		sort.Slice(instances, func(i, j int) bool {
+			return instances[i].State == "ACTIVE"
+		})
+
 		for _, job := range queue {
 			if job.State == autoscale.RUNNING {
 				runningJobs++
 				continue
 			}
-			instances, err := curClust.GetInstances()
+			instances, err = curClust.GetInstances()
 			if err != nil {
-				return out, err
+				return autoscale.AlgorithmOutput{}, err
 			}
 			//Check if there are room to add more cluster
 			if curClust.GetInstanceLimit() > len(instances) {
@@ -123,6 +134,15 @@ func (n NaiveAlgorithm) Run(input autoscale.AlgorithmInput, startTime time.Time)
 					Type:  iType.Name,
 					State: "",
 				}
+
+				for k := activeInstances; k < len(instances); k++ {
+					if instances[k].State == "INACTIVE" {
+						instance = instances[k]
+						activeInstances++
+						break
+					}
+				}
+
 				_, err = curClust.AddInstance(&instance, startTime)
 				if err != nil {
 					return autoscale.AlgorithmOutput{}, err
@@ -130,12 +150,10 @@ func (n NaiveAlgorithm) Run(input autoscale.AlgorithmInput, startTime time.Time)
 				outInstances = append(outInstances, instance)
 			} else {
 				//If all instances have been used, check for inactive instances and use them instead
-				for _, i := range instances {
-					if i.State == autoscale.INACTIVE {
-						_, err = curClust.AddInstance(&i, startTime)
-						if err != nil {
-							return autoscale.AlgorithmOutput{}, err
-						}
+				for k := activeInstances-1; k < len(instances); k++ {
+					if instances[k].State == "INACTIVE" {
+						_, err = curClust.AddInstance(&instances[k], startTime)
+						activeInstances++
 						break
 					}
 				}
@@ -143,17 +161,17 @@ func (n NaiveAlgorithm) Run(input autoscale.AlgorithmInput, startTime time.Time)
 		}
 		//DELETE instances based on the queue size, if the instance size is larger than queue,
 		//there should be at least one INACTIVE instance
-		if len(instances) > (len(queue)) {
-			for _, i := range instances {
-				if i.State == autoscale.INACTIVE {
-					err = curClust.DeleteInstance(i.Id, startTime)
-					if err != nil {
-						return autoscale.AlgorithmOutput{}, err
-					}
-					outInstances = append(outInstances, i)
-				}
-			}
-		}
+		//if len(instances) > (len(queue)) {
+		//	for _, i := range instances {
+		//		if i.State == autoscale.INACTIVE {
+		//			err = curClust.DeleteInstance(i.Id, startTime)
+		//			if err != nil {
+		//				return autoscale.AlgorithmOutput{}, err
+		//			}
+		//			outInstances = append(outInstances, i)
+		//		}
+		//	}
+		//}
 
 		for key, _ := range input.Clouds {
 			if _, ok := queueMap[key]; !ok {
@@ -161,11 +179,13 @@ func (n NaiveAlgorithm) Run(input autoscale.AlgorithmInput, startTime time.Time)
 				if err != nil {
 					return autoscale.AlgorithmOutput{}, err
 				}
-				for _, inst := range instances {
-					if inst.State == autoscale.INACTIVE {
-						err := input.Clouds[key].DeleteInstance(inst.Id, startTime)
-						if err != nil {
-							return autoscale.AlgorithmOutput{}, err
+				if len(instances) > (len(queue)) {
+					for _, inst := range instances {
+						if inst.State == autoscale.INACTIVE {
+							err := input.Clouds[key].DeleteInstance(inst.Id, startTime)
+							if err != nil {
+								return autoscale.AlgorithmOutput{}, err
+							}
 						}
 					}
 				}
